app now successfully redirects when user logs in
