+++ screens/Login/index.jsx
@@ -11,10 +11,12 @@ import React, { useState } from "react";
 import { TextInput as MaterialTI } from "react-native-paper";
 import { useNavigation } from "@react-navigation/native";
 import { addObjectToCache } from "../../utiils/caching";
+import { useGlobalContext } from "../../context";
 
 const Login = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const { setUser } = useGlobalContext();
   const navigation = useNavigation();
 
   const handleLogin = async (email, password) => {
@@ -26,6 +28,7 @@ const Login = () => {
       });
       console.log("response: ", response.data);
       addObjectToCache("user", response.data.user);
+      setUser(response.data.user);
     } catch (error) {
       console.error(error.message);
     }
